wip: debugging 16-bit simd128

+++ libcrux-ml-kem/src/simd/portable.rs
@@ -71,7 +71,7 @@ fn bitwise_and_with_constant(mut v: PortableVector, c: i16) -> PortableVector {
 }
 
 #[inline(always)]
-fn shift_right<const SHIFT_BY: i16>(mut v: PortableVector) -> PortableVector {
+fn shift_right<const SHIFT_BY: i32>(mut v: PortableVector) -> PortableVector {
     for i in 0..FIELD_ELEMENTS_IN_VECTOR {
         v.elements[i] = v.elements[i] >> SHIFT_BY;
     }
@@ -80,7 +80,7 @@ fn shift_right<const SHIFT_BY: i16>(mut v: PortableVector) -> PortableVector {
 }
 
 #[inline(always)]
-fn shift_left<const SHIFT_BY: i16>(mut lhs: PortableVector) -> PortableVector {
+fn shift_left<const SHIFT_BY: i32>(mut lhs: PortableVector) -> PortableVector {
     for i in 0..FIELD_ELEMENTS_IN_VECTOR {
         lhs.elements[i] = lhs.elements[i] << SHIFT_BY;
     }
@@ -135,7 +135,7 @@ fn compress_1(mut v: PortableVector) -> PortableVector {
 }
 
 #[inline(always)]
-fn compress<const COEFFICIENT_BITS: i16>(mut v: PortableVector) -> PortableVector {
+fn compress<const COEFFICIENT_BITS: i32>(mut v: PortableVector) -> PortableVector {
     for i in 0..FIELD_ELEMENTS_IN_VECTOR {
         v.elements[i] =
             compress_ciphertext_coefficient(COEFFICIENT_BITS as u8, v.elements[i] as u16) as i16;
@@ -143,7 +143,7 @@ fn compress<const COEFFICIENT_BITS: i16>(mut v: PortableVector) -> PortableVecto
     v
 }
 
-fn decompress<const COEFFICIENT_BITS: i16>(mut v: PortableVector) -> PortableVector {
+fn decompress<const COEFFICIENT_BITS: i32>(mut v: PortableVector) -> PortableVector {
     debug_assert!(to_i16_array(v)
         .into_iter()
         .all(|coefficient| coefficient.abs() < 1 << COEFFICIENT_BITS));
@@ -526,11 +526,11 @@ impl Operations for PortableVector {
         bitwise_and_with_constant(v, c)
     }
 
-    fn shift_right<const SHIFT_BY: i16>(v: Self) -> Self {
+    fn shift_right<const SHIFT_BY: i32>(v: Self) -> Self {
         shift_right::<{ SHIFT_BY }>(v)
     }
 
-    fn shift_left<const SHIFT_BY: i16>(v: Self) -> Self {
+    fn shift_left<const SHIFT_BY: i32>(v: Self) -> Self {
         shift_left::<{ SHIFT_BY }>(v)
     }
 
@@ -550,11 +550,11 @@ impl Operations for PortableVector {
         compress_1(v)
     }
 
-    fn decompress<const COEFFICIENT_BITS: i16>(v: Self) -> Self {
+    fn decompress<const COEFFICIENT_BITS: i32>(v: Self) -> Self {
         decompress::<COEFFICIENT_BITS>(v)
     }
 
-    fn compress<const COEFFICIENT_BITS: i16>(v: Self) -> Self {
+    fn compress<const COEFFICIENT_BITS: i32>(v: Self) -> Self {
         compress::<COEFFICIENT_BITS>(v)
     }
